% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_FEM.R
\name{Compute_B_sB_old}
\alias{Compute_B_sB_old}
\title{Create the Design matrix of the sB approximation (without the B spline)}
\usage{
Compute_B_sB_old(x, a, k, region, boundary = TRUE)
}
\arguments{
\item{x}{A vector of numeric values that denotes the evaluation locations.}

\item{a}{A positive scalar represents the periodicity parameter.}

\item{k}{A positive integer represents the number of knots used to define the sB basis. The number of
basis functions equals to 2 times k or 2 times (k-1) if boundary is TRUE.}

\item{region}{A vector of size 2 that denotes the upper and lower interval limit of the region of interest.}

\item{boundary}{A logical value to denote whether to consider the boundary conditions.}
}
\value{
A sparse matrix as the design matrix of the weights.
}
\description{
This function outputs the design matrix of the sB approximation as a sparse matrix, given the number of knots,
the periodicity parameter a, and the region of interest.
}
\examples{
Compute_B_sB_old(x = c(1,2,3), a = 1, k = 5, region = c(0,3))
}
