% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_Prior.R
\name{compute_d_step_sGPsd}
\alias{compute_d_step_sGPsd}
\title{Compute the correction factor based on the PSD.}
\usage{
compute_d_step_sGPsd(d, a)
}
\arguments{
\item{d}{A positive scalar represents the prediction unit.}

\item{a}{A positive scalar represents the frequency parameter.}
}
\value{
A scalar denotes the correction factor to recover the original SD.
}
\description{
This function outputs the correction factor that converts the prior on
PSD to the prior on the original SD, based on the prediction unit d and
frequency parameter a.
}
