% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_trueGP.R
\name{compute_matrix_given_cov}
\alias{compute_matrix_given_cov}
\title{Compute covariance matrix given covariance function.}
\usage{
compute_matrix_given_cov(from, to, m, K)
}
\arguments{
\item{from}{numeric value for the left limit of the interval of interest}

\item{to}{numeric value for the right limit of the interval of interest}

\item{m}{An integer value of number of points to simulate.}

\item{K}{A covariance function of two arguments that is PD.}
}
\value{
A m by m covariance matrix
}
\description{
This function computes the covariance matrix of the GP given its covariance function,
over a specified region.
}
\examples{
fun <- generate_K_true(1,1)
compute_matrix_given_cov(from = 0, to = 1, m = 5, K = fun)
}
