% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_FEM.R
\name{Compute_Q_sB}
\alias{Compute_Q_sB}
\title{Create the Precision matrix of the sB approximation}
\usage{
Compute_Q_sB(a, k, region, accuracy = 0.01, boundary = TRUE)
}
\arguments{
\item{a}{A positive scalar represents the periodicity parameter.}

\item{k}{A positive integer represents the number of knots used to define the sB basis. The number of
basis functions equals to 2 times k or 2 times (k-1) if boundary is TRUE.}

\item{region}{A vector of size 2 that denotes the upper and lower interval limit of the region of interest.}

\item{accuracy}{A positive integer represents the integration size used to compute the inner product. Smaller value
means more accurate inner product with longer computational time.}

\item{boundary}{A logical value to denote whether to consider the boundary conditions.}
}
\value{
A sparse matrix as the precision matrix of the weights.
}
\description{
This function outputs the precision matrix of the sB approximation as a sparse matrix, given the number of knots,
the periodicity parameter a, and the region of interest.
}
\examples{
Compute_Q_sB(a = 1, k = 5, region = c(0,1))
}
