% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_trueGP.R
\name{gaussprocess}
\alias{gaussprocess}
\title{Simulates a Gaussian process with a given covariance function.}
\usage{
gaussprocess(
  from = 0,
  to = 1,
  K = function(s, t) {     min(s, t) },
  start = 0,
  m = 1000
)
}
\arguments{
\item{from}{numeric value for the left limit of the interval of interest}

\item{to}{numeric value for the right limit of the interval of interest}

\item{K}{A covariance function of two arguments that is PD. Default being Brownian motion.}

\item{start}{numerical of the starting value of the GP, default to zero.}

\item{m}{An integer value of number of points to simulate.}
}
\value{
A dataframe with time index and value of the sample path.
}
\description{
This function simulates a Gaussian process over a fine-grid,
provided with a given covariance function.
}
\examples{
gaussprocess(from = 0, to = 1, K = function(s, t) {min(s, t)}, start = 0, m = 10)
}
