% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_trueGP.R
\name{generate_K_true}
\alias{generate_K_true}
\title{Generate the true covariance function of the sGP, given its periodicity parameter
and its SD parameter.}
\usage{
generate_K_true(sigma, alpha)
}
\arguments{
\item{sigma}{The value of the SD parameter.}

\item{alpha}{The value of the periodicity parameter.}
}
\value{
A PD covariance function that takes two arguments.
}
\description{
This function takes the periodicity parameter and SD parameter to generate a covariance function
that correspond to the specified sGP.
}
\examples{
fun <- generate_K_true(1,1)
fun(1,2)
}
