% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_StateSpace.R
\name{joint_prec_construct}
\alias{joint_prec_construct}
\title{Construct the joint precision matrix of the sGP process and its derivative.}
\usage{
joint_prec_construct(t_vec, a, sd)
}
\arguments{
\item{t_vec}{A positive numeric vector denotes the locations to evaluate the GP.}

\item{a}{A positive scalar represents the frequency parameter.}

\item{sd}{A positive scalar represents the standard deviation parameter.}
}
\value{
A sparse precision matrix for the sGP (and derivative) at the locations.
}
\description{
This function constructs the joint precision matrix of the sGP process and its
derivative (second entry) given the sGP's frequency and standard deviation,
at a given set of locations.
}
\examples{
joint_prec_construct(t_vec = c(1,2,3), a = 1, sd = 1)
}
