% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_trueGP.R
\name{sim_GP}
\alias{sim_GP}
\title{Simulates a Gaussian process with a given covariance function.}
\usage{
sim_GP(
  grid,
  K = function(s, t) {
     min(s, t)
 }
)
}
\arguments{
\item{grid}{A vector of locations to be provided as the grid.}

\item{K}{A covariance function of two arguments that is PD. Default being Brownian motion.}
}
\value{
A dataframe with time index and value of the sample path.
}
\description{
This function simulates a Gaussian process over a fine-grid,
provided with a given covariance function.
}
\examples{
sim_GP(grid = c(0.1,0.2,0.3), K = function(s, t) {min(s, t)})
}
