% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_StateSpace.R
\name{sim_sGP_Var}
\alias{sim_sGP_Var}
\title{Simulate the exact sGP and its derivative at a fine mesh, using the state-space
representation.}
\usage{
sim_sGP_Var(t = NULL, mesh_size = 0.01, max_t = 10, a, sd)
}
\arguments{
\item{t}{A vector of locations to evaluate the sGP and its derivative, by default being NULL.}

\item{mesh_size}{A scalar denotes the size of each mesh to create, if t is NULL}

\item{max_t}{A scalar denotes the maximum of location to consider if t is NULL.}

\item{a}{A positive scalar represents the periodicity parameter.}

\item{sd}{A positive scalar represents the standard deviation parameter.}
}
\value{
A dataframe that contains the sampled sGP and its derivative at the vector t.
}
\description{
This function simulates sample path of the exact sGP with its derivatives at a fine
mesh, using the state-space representation.
}
