% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_trueGP.R
\name{compute_matrix_given_cov}
\alias{compute_matrix_given_cov}
\title{Compute covariance matrix given covariance function.}
\usage{
compute_matrix_given_cov(grid, K)
}
\arguments{
\item{grid}{A vector of locations to be provided as the grid.}

\item{K}{A covariance function of two arguments that is PD.}
}
\value{
A m by m covariance matrix, with m being length of grid.
}
\description{
This function computes the covariance matrix of the GP given its covariance function,
over a specified region.
}
\examples{
fun <- generate_K_true(1,1)
compute_matrix_given_cov(grid = c(0.1,0.2,0.3,0.4), K = fun)
}
